Probably fix playglobalsound completion
The path completion works locally but doesn't work on live. NFI when this broke but if this doesn't fix I'll bisect and dig (it's probably respath?)
