+++ Content.Server/Audio/ServerGlobalSoundSystem.cs
@@ -1,9 +1,5 @@
-﻿using Content.Server.Administration;
-using Content.Server.Station.Components;
-using Content.Server.Station.Systems;
-using Content.Shared.Administration;
+﻿using Content.Server.Station.Systems;
 using Content.Shared.Audio;
-using Robust.Server.Player;
 using Robust.Shared.Audio;
 using Robust.Shared.Console;
 using Robust.Shared.Player;
@@ -13,14 +9,7 @@ namespace Content.Server.Audio;
 public sealed class ServerGlobalSoundSystem : SharedGlobalSoundSystem
 {
     [Dependency] private readonly IConsoleHost _conHost = default!;
-    [Dependency] private readonly IPlayerManager _playerManager = default!;
     [Dependency] private readonly StationSystem _stationSystem = default!;
-
-    public override void Initialize()
-    {
-        base.Initialize();
-        _conHost.RegisterCommand("playglobalsound", Loc.GetString("play-global-sound-command-description"), Loc.GetString("play-global-sound-command-help"), PlayGlobalSoundCommand);
-    }
 
     public override void Shutdown()
     {
@@ -28,7 +17,7 @@ public sealed class ServerGlobalSoundSystem : SharedGlobalSoundSystem
         _conHost.UnregisterCommand("playglobalsound");
     }
 
-    private void PlayAdminGlobal(Filter playerFilter, string filename, AudioParams? audioParams = null, bool replay = true)
+    public void PlayAdminGlobal(Filter playerFilter, string filename, AudioParams? audioParams = null, bool replay = true)
     {
         var msg = new AdminSoundEvent(filename, audioParams);
         RaiseNetworkEvent(msg, playerFilter, recordReplay: replay);
@@ -67,79 +56,5 @@ public sealed class ServerGlobalSoundSystem : SharedGlobalSoundSystem
 
         var filter = GetStationAndPvs(source);
         RaiseNetworkEvent(msg, filter);
-    }
-
-    /// <summary>
-    ///     Command that allows admins to play global sounds.
-    /// </summary>
-    [AdminCommand(AdminFlags.Fun)]
-    public void PlayGlobalSoundCommand(IConsoleShell shell, string argStr, string[] args)
-    {
-        Filter filter;
-        var audio = AudioParams.Default;
-
-        bool replay = true;
-
-        switch (args.Length)
-        {
-            // No arguments, show command help.
-            case 0:
-                shell.WriteLine(Loc.GetString("play-global-sound-command-help"));
-                return;
-
-            // No users, play sound for everyone.
-            case 1:
-                // Filter.Broadcast does resolves IPlayerManager, so use this instead.
-                filter = Filter.Empty().AddAllPlayers(_playerManager);
-                break;
-
-            // One or more users specified.
-            default:
-                var volumeOffset = 0;
-
-                // Try to specify a new volume to play it at.
-                if (int.TryParse(args[1], out var volume))
-                {
-                    audio = audio.WithVolume(volume);
-                    volumeOffset = 1;
-                }
-                else
-                {
-                    shell.WriteError(Loc.GetString("play-global-sound-command-volume-parse", ("volume", args[1])));
-                    return;
-                }
-
-                // No users specified so play for them all.
-                if (args.Length == 2)
-                {
-                    filter = Filter.Empty().AddAllPlayers(_playerManager);
-                }
-                else
-                {
-                    // TODO REPLAYS uhhh.. what to do with this?
-                    replay = false;
-
-                    filter = Filter.Empty();
-
-                    // Skip the first argument, which is the sound path.
-                    for (var i = 1 + volumeOffset; i < args.Length; i++)
-                    {
-                        var username = args[i];
-
-                        if (!_playerManager.TryGetSessionByUsername(username, out var session))
-                        {
-                            shell.WriteError(Loc.GetString("play-global-sound-command-player-not-found", ("username", username)));
-                            continue;
-                        }
-
-                        filter.AddPlayer(session);
-                    }
-                }
-
-                break;
-        }
-
-        audio = audio.AddVolume(-8);
-        PlayAdminGlobal(filter, args[0], audio, replay);
     }
 }
